Move LinkContactCommandTest to correct location
