+++ src/test/java/seedu/address/logic/parser/LinkContactCommandParserTest.java
@@ -1,0 +1,2 @@
+package seedu.address.logic.parser;public class LinkContactCommandParserTest {
+}
